Iran throttling case
added the following language to the HTTP case: While this is commendable, we must not lose track of the fact that different protocols, implementations and networking paradigms can intersect such that they adversely impact human rights. For instance, certain countries will throttle HTTPS connections forcing users to switch to the (unthrottled) HTTP to facilitating surveillance. {{MITtechnologyreview2013}}.